removes unnecessary method in rabbit configuration
